ДЗ 5.15. Модуляция
