+++ js/utils/choose.js
@@ -1,0 +1,4 @@
+export function choose(choices) {
+  const index = Math.floor(Math.random() * choices.length);
+  return choices[index];
+}
