+++ js/utils/getCommentNumberMargin.js
@@ -1,0 +1,1 @@
+export const getCommentNumberMargin = (i) => (i - 1) * 5;
